requirements.txt
streamlit==1.31.0
pandas==2.2.3
opencv-python==4.10.0.84
Pillow==10.4.0
ultralytics==8.3.22
python-dotenv==1.0.1
langchain==0.3.4
langchain-groq==0.2.0


